oforha.ai/projects/templates/requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning
scikit-learn>=0.24.0
tensorflow>=2.6.0
tensorflow-js>=3.9.0

# Natural Language Processing
spacy>=3.1.0
nltk>=3.6.0
transformers>=4.11.0
sentence-transformers>=2.2.0
word2vec>=0.11.0

# RAG Components
faiss-cpu>=1.7.0
chromadb>=0.3.0
langchain>=0.0.150
elasticsearch>=7.14.0

# Web Development
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0

# Database
pymongo>=3.12.0
redis>=3.5.0
sqlalchemy>=1.4.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0
black>=21.7b0
pylint>=2.9.0
pytest>=6.2.0

# Monitoring
prometheus-client>=0.11.0
grafana-api>=1.0.0

# Utilities
python-dotenv>=0.19.0
requests>=2.26.0
tqdm>=4.62.0
pyyaml>=5.4.0
optuna>=2.10.0
mlflow>=1.20.0
joblib>=1.0.0 